+++ ExpenseProcessingSystem/ExpenseProcessingSystem/Models/BIR2307/PaymentInfo.cs
@@ -9,16 +9,16 @@ namespace BIR_Form_Filler.Models
     {
         private string payments;
         private string atc;
-        private double m1Quarter;
-        private double m2Quarter;
-        private double m3Quarter;
-        private double taxWithheld;
+        private decimal m1Quarter;
+        private decimal m2Quarter;
+        private decimal m3Quarter;
+        private decimal taxWithheld;
 
         public string Payments { get => payments; set => payments = value; }
         public string Atc { get => atc; set => atc = value; }
-        public double M1Quarter { get => m1Quarter; set => m1Quarter = value; }
-        public double M2Quarter { get => m2Quarter; set => m2Quarter = value; }
-        public double M3Quarter { get => m3Quarter; set => m3Quarter = value; }
-        public double TaxWithheld { get => taxWithheld; set => taxWithheld = value; }
+        public decimal M1Quarter { get => m1Quarter; set => m1Quarter = value; }
+        public decimal M2Quarter { get => m2Quarter; set => m2Quarter = value; }
+        public decimal M3Quarter { get => m3Quarter; set => m3Quarter = value; }
+        public decimal TaxWithheld { get => taxWithheld; set => taxWithheld = value; }
     }
 }
