+++ ExpenseProcessingSystem/ExpenseProcessingSystem/Models/ExpenseEntryCashBreakdownModel.cs
@@ -11,7 +11,7 @@ namespace ExpenseProcessingSystem.Models
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int CashBreak_ID { get; set; }
-        public double CashBreak_Denomination { get; set; }
+        public decimal CashBreak_Denomination { get; set; }
         public int CashBreak_NoPcs { get; set; }
         public decimal CashBreak_Amount { get; set; }
         public ExpenseEntryDetailModel ExpenseEntryDetailModel { get; set; }
