+++ ExpenseProcessingSystem/ExpenseProcessingSystem/ViewModels/Entry/DDVInterEntityViewModel.cs
@@ -11,10 +11,10 @@ namespace ExpenseProcessingSystem.ViewModels.Entry
         public int Inter_Currency2_ID { get; set; }
         public string Inter_Currency1_ABBR { get; set; }
         public string Inter_Currency2_ABBR { get; set; }
-        public float Inter_Currency1_Amount { get; set; }
-        public float Inter_Currency2_Amount { get; set; }
-        public float Inter_Convert1_Amount { get; set; }
-        public float Inter_Convert2_Amount { get; set; }
+        public decimal Inter_Currency1_Amount { get; set; }
+        public decimal Inter_Currency2_Amount { get; set; }
+        public decimal Inter_Convert1_Amount { get; set; }
+        public decimal Inter_Convert2_Amount { get; set; }
         public bool Inter_Check1 { get; set; }
         public bool Inter_Check2 { get; set; }
         public float Inter_Rate { get; set; }
@@ -46,18 +46,18 @@ namespace ExpenseProcessingSystem.ViewModels.Entry
     {
         public int Inter_Acc_ID { get; set; }
         public int Inter_Curr_ID { get; set; }
-        public float Inter_Amount { get; set; }
-        public float Inter_Rate { get; set; }
+        public decimal Inter_Amount { get; set; }
+        public decimal Inter_Rate { get; set; }
         public int Inter_Type_ID { get; set; }
     }
     public class InterEntityParticular
     {
         public int Particular_Acc_ID { get; set; }
         public string Particular_Account_Name { get; set; }
-        public float Particular_Debit_Amount { get; set; }
+        public decimal Particular_Debit_Amount { get; set; }
         public int Particular_DebCurr_ID { get; set; }
         public string Particular_Debit_Curr { get; set; }
-        public float Particular_Credit_Amount { get; set; }
+        public decimal Particular_Credit_Amount { get; set; }
         public int Particular_CredCurr_ID { get; set; }
         public string Particular_Credit_Curr { get; set; }
         public float Particular_Credit_Rate { get; set; }
